Add file-based run-time configuration.
Instead of requiring users to recompile kcaldav for debugging (in,
say, the event of errors in parsing or changes to clients), use a
run-time configuration file by default in /conf/kcaldav.conf.

If this file does not exist, or is not specified during compile
time under CFGFILE, then use the defaults.

Continue to accept LOGFILE and DEBUG, but consider these to be
deprecated.  They are overridden by the elements in kcaldav.conf.