Fix RRULE UNTIL conditional date format.
As specified in RFC 5545, pg. 41, the UNTIl statement of a RRULE
must match the time-zone configuration of its context.

This change first introduces a "type" to "struct icaltm" that
indicates the type of a time-stamp: UTC, date, or date-time.  With
this information, a component or sub-component can verify that the
UNTIL statement matches.

Matching occurs by seeing if the context specifies time-zone
information with its date-time (thus making an absolute time) or
outright uses UTC.  If so, the UNTIL must be UTC.  Otherwise,
the UNTIL must inherit.

This is from a bug report by Stefan Sperling.  Thank you!